requirements.txt
loguru==0.6.0
pyOpenSSL==22.0.0